requirements.txt
PyQt6
PyQt6-WebEngine
spotipy
platformdirs
python-dotenv
